backend/pyproject.toml
[project]
name = "tasktracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.3",
    "django-elasticsearch-dsl>=8.2",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "envparse>=0.2.0",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]
